+++ src/main/java/com/cpi/is/service/impl/maintenance/MaterialCodeServiceImpl.java
@@ -36,14 +36,45 @@ public class MaterialCodeServiceImpl implements MaterialCodeService {
 
 	@Override
 	public String saveData(HttpServletRequest request) throws Exception {
-		return materialCodeDAO.saveData(
-				jsonToEntity(new JSONObject(request.getParameter("data"))));
+		String validation = validateData(request);
+		if(validation.equals("Success")) {
+			return materialCodeDAO.saveData(
+					jsonToEntity(new JSONObject(request.getParameter("data"))));
+		} else {
+			return validation;
+		}
 	}
 
 	@Override
 	public String deleteData(HttpServletRequest request) throws Exception {
-		return materialCodeDAO.deleteData(
-				jsonToEntity(new JSONObject(request.getParameter("data"))));
+		String validation = validateData(request);
+		if(validation.equals("Success")) {
+			return materialCodeDAO.deleteData(
+					jsonToEntity(new JSONObject(request.getParameter("data"))));
+		} else {
+			return "Unable to delete data";
+		}
+	}
+	
+	public String validateData(HttpServletRequest request) throws Exception{
+		JSONObject json = new JSONObject(request.getParameter("data"));
+		String validation = "Success";
+		
+		if (!json.has("materialCd") || !(json.get("materialCd") instanceof String)) {
+			validation = "Please fill-out the material code form properly";
+		} else if (!json.has("materialCodeName") || !(json.get("materialCodeName") instanceof String)) {
+			validation = "Please fill-out the material code form properly";
+		} else if (!json.has("unitOfMeasurement") || !(json.get("unitOfMeasurement") instanceof String)) {
+			validation = "Please fill-out the material code form properly";
+		} else if (json.getString("materialCd").length() < 1 || json.getString("materialCd").length() > 50) {
+			validation = "Please fill-out the material code form properly";
+		} else if (json.getString("materialCodeName").length() < 1 || json.getString("materialCodeName").length() > 200) {
+			validation = "Please fill-out the material code form properly";
+		} else if (json.getString("unitOfMeasurement").length() < 1 || json.getString("unitOfMeasurement").length() > 100) {
+			validation = "Please fill-out the material code form properly";
+		}
+		
+		return validation;
 	}
 
 }
